app/backend/requirements.txt
flask
psycopg2-binary
prometheus_client